factorize NAL filtering, vaapi->sw fallback
